Update the jsonl script
